Fix type of 'configs' property
Correct the definition of the 'configs' property in the JSON
returned by listing topics or getting details of a topic. Previously
it was incorrectly described as being an array of JSON objects, when
in fact it is a JSON object. Also add in some properties of the
`configs` object not previously included in the OpenAPI
specification.